Put enums into their own files and make them distinct classes
