+++ nephthys/macros/__init__.py
@@ -1,7 +1,7 @@
 from typing import Any
 
+from nephthys.database.enums import TicketStatus
 from nephthys.database.tables import Ticket
-from nephthys.database.tables import TicketStatus
 from nephthys.database.tables import User
 from nephthys.macros.faq import FAQ
 from nephthys.macros.fraud import Fraud
