+++ nephthys/macros/thread.py
@@ -1,7 +1,7 @@
 from slack_sdk.errors import SlackApiError
 
 from nephthys.actions.resolve import resolve
-from nephthys.database.tables import TicketStatus
+from nephthys.database.enums import TicketStatus
 from nephthys.database.tables import User
 from nephthys.macros.types import Macro
 from nephthys.utils.env import env
