+++ nephthys/piccolo_app.py
@@ -1,8 +1,8 @@
 import os
 
 from piccolo.conf.apps import AppConfig
-from piccolo.conf.apps import get_package
-from piccolo.conf.apps import table_finder
+
+from nephthys.database.tables import ALL_TABLES
 
 CURRENT_DIRECTORY = os.path.dirname(os.path.abspath(__file__))
 
@@ -10,11 +10,7 @@ CURRENT_DIRECTORY = os.path.dirname(os.path.abspath(__file__))
 APP_CONFIG = AppConfig(
     app_name="nephthys",
     migrations_folder_path=os.path.join(CURRENT_DIRECTORY, "piccolo_migrations"),
-    table_classes=table_finder(
-        modules=["nephthys.database.tables"],
-        package=get_package(__name__),
-        exclude_imported=True,
-    ),
+    table_classes=ALL_TABLES,
     migration_dependencies=[],
     commands=[],
 )
